mogilefs: all integers are long integers

+++ src/collectors/mogilefs/mogilefs.py
@@ -41,7 +41,7 @@ class MogilefsCollector(diamond.collector.Collector):
 
         for line in out.splitlines()[:-1]:
             name, var = line.partition(" ")[::2]
-            myvars[name.strip()] = long(var)
+            myvars[name.strip()] = var
 
         for key, value in myvars.iteritems():
             # Set Metric Name
